MIR: use layout info for tuple aggregates

Incorporates https://github.com/GaloisInc/mir-json/pull/218 and https://github.com/GaloisInc/crucible/pull/1743.  This also required a small tweak to the code for building `TupleShape`s (but not to anything else, since all the other tuple code uses the offsets from the `TupleShape` instead of computing them directly).

We actually had no test coverage for this change, since every crux-mir-comp test case that uses tuples happened to use `(u8, u8)`, where the real offsets are the same as the previous "hardcoded size=1" offsets.  I've changed a few of these to use `(u16, u16)` instead, and confirmed that they do fail without the `TupleShape` update.

+++ crux-mir-comp/test/symb_eval/comp/spec_tuple.rs
@@ -2,28 +2,28 @@ extern crate crucible;
 use crucible::*;
 use crucible::method_spec::{MethodSpec, MethodSpecBuilder, clobber_globals};
 
-fn f(x: (u8, u8)) -> (u8, u8) {
+fn f(x: (u16, u16)) -> (u16, u16) {
     (x.1, x.0)
 }
 
 #[crux::test]
 fn f_test() {
     clobber_globals();
-    let x = <(u8, u8)>::symbolic("x");
+    let x = <(u16, u16)>::symbolic("x");
     crucible_assume!(x.0 > 0);
     let y = f(x);
     crucible_assert!(y.1 > 0);
 }
 
 fn f_spec() -> MethodSpec {
-    let x = <(u8, u8)>::symbolic("x");
+    let x = <(u16, u16)>::symbolic("x");
     crucible_assume!(x.0 > 0);
 
     let mut msb = MethodSpecBuilder::new(f);
     msb.add_arg(&x);
     msb.gather_assumes();
 
-    let y = <(u8, u8)>::symbolic("result");
+    let y = <(u16, u16)>::symbolic("result");
     crucible_assert!(y.1 > 0);
 
     msb.set_return(&y);
@@ -47,8 +47,8 @@ fn f_spec() -> MethodSpec {
 fn use_f() {
     f_spec().enable();
 
-    let a = u8::symbolic("a");
-    let b = u8::symbolic("b");
+    let a = u16::symbolic("a");
+    let b = u16::symbolic("b");
     crucible_assume!(0 < a && a < 10);
     let (c, d) = f((a, b));
     crucible_assert!(0 < d);
