+++ crux-mir-comp/test/symb_eval/cryptol/basic_err_tuple_size.rs
@@ -5,5 +5,5 @@ const PATH: &str = "test::symb_eval::cryptol::basic";
 
 #[crux::test]
 fn test() {
-    let f: fn((u8, u8, u8)) -> u8 = cryptol::load(PATH, "tupleArg");
+    let f: fn((u16, u16, u16)) -> u16 = cryptol::load(PATH, "tupleArg");
 }
